+++ packages/markdown-overrides/babel.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  extends: '../../babel.defaults.js',
-  presets: ['codecademy', '@babel/preset-typescript'],
-};