+++ packages/monaco/jest.config.ts
@@ -1,5 +1,0 @@
-import base from '../../jest.config.base';
-
-export default base('monaco', {
-  transformIgnorePatterns: ['node_modules'],
-});