+++ packages/styleguide/.storybook/main.ts
@@ -23,7 +23,7 @@ module.exports = {
   ],
   stories: ['../stories/**/*.stories.@(mdx|tsx)'],
   typescript: {
-    reactDocgen: 'react-docgen-typescript',
+    reactDocgen: 'react-docgen-typescript-plugin',
     reactDocgenTypescriptOptions: {
       shouldExtractLiteralValuesFromEnum: true,
       shouldRemoveUndefinedFromOptional: true,
@@ -80,10 +80,6 @@ module.exports = {
         ),
 
         '@codecademy/gamut$': path.resolve(__dirname, '../../gamut/src'),
-        '@codecademy/gamut-labs$': path.resolve(
-          __dirname,
-          '../../gamut-labs/src'
-        ),
         '@codecademy/gamut-illustrations$': path.resolve(
           __dirname,
           '../../gamut-illustrations/src'