+++ packages/gamut-icons/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [9.3.4](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.3.3...@codecademy/gamut-icons@9.3.4) (2023-07-05)
+
+**Note:** Version bump only for package @codecademy/gamut-icons
+
+
+
+
+
+### [9.3.3](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.3.2...@codecademy/gamut-icons@9.3.3) (2023-06-21)
+
+**Note:** Version bump only for package @codecademy/gamut-icons
+
+
+
+
+
 ### [9.3.2](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.3.1...@codecademy/gamut-icons@9.3.2) (2023-05-31)
 
 **Note:** Version bump only for package @codecademy/gamut-icons