+++ packages/gamut-kit/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.6.335](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.334...@codecademy/gamut-kit@0.6.335) (2023-07-27)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.6.334](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.333...@codecademy/gamut-kit@0.6.334) (2023-07-20)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.6.333](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.332...@codecademy/gamut-kit@0.6.333) (2023-07-13)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.6.332](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.331...@codecademy/gamut-kit@0.6.332) (2023-07-12)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.6.331](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.330...@codecademy/gamut-kit@0.6.331) (2023-07-12)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.6.330](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.329...@codecademy/gamut-kit@0.6.330) (2023-07-05)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.6.329](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.328...@codecademy/gamut-kit@0.6.329) (2023-06-21)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.6.328](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.327...@codecademy/gamut-kit@0.6.328) (2023-06-15)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
 ### [0.6.327](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.326...@codecademy/gamut-kit@0.6.327) (2023-06-05)
 
 **Note:** Version bump only for package @codecademy/gamut-kit