+++ packages/gamut-labs/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [35.15.8](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-labs@35.15.7...@codecademy/gamut-labs@35.15.8) (2023-07-27)
+
+**Note:** Version bump only for package @codecademy/gamut-labs
+
+
+
+
+
+### [35.15.7](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-labs@35.15.6...@codecademy/gamut-labs@35.15.7) (2023-07-20)
+
+**Note:** Version bump only for package @codecademy/gamut-labs
+
+
+
+
+
+### [35.15.6](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-labs@35.15.5...@codecademy/gamut-labs@35.15.6) (2023-07-13)
+
+**Note:** Version bump only for package @codecademy/gamut-labs
+
+
+
+
+
+### [35.15.5](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-labs@35.15.4...@codecademy/gamut-labs@35.15.5) (2023-07-12)
+
+**Note:** Version bump only for package @codecademy/gamut-labs
+
+
+
+
+
+### [35.15.4](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-labs@35.15.3...@codecademy/gamut-labs@35.15.4) (2023-07-12)
+
+**Note:** Version bump only for package @codecademy/gamut-labs
+
+
+
+
+
+### [35.15.3](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-labs@35.15.2...@codecademy/gamut-labs@35.15.3) (2023-07-05)
+
+**Note:** Version bump only for package @codecademy/gamut-labs
+
+
+
+
+
+### [35.15.2](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-labs@35.15.1...@codecademy/gamut-labs@35.15.2) (2023-06-21)
+
+**Note:** Version bump only for package @codecademy/gamut-labs
+
+
+
+
+
 ### [35.15.1](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-labs@35.15.0...@codecademy/gamut-labs@35.15.1) (2023-06-05)
 
 **Note:** Version bump only for package @codecademy/gamut-labs